feat(orders): refactor order display and payment status handling, add OrderCard and OrderProgress components
